Only allow destroying the most recent result. Issue #8.

+++ app/services/result_service.rb
@@ -28,12 +28,16 @@ class ResultService
   end
 
   def self.destroy(result)
+    return OpenStruct.new(:success? => false) unless result.most_recent?
+
     Result.transaction do
       [result.winner, result.loser].each do |player|
         player.rewind_rating!(result.game)
       end
 
       result.destroy
+
+      OpenStruct.new(:success? => true)
     end
   end
 end
